+++ lib/active_model/serializer.rb
@@ -1,7 +1,7 @@
 require 'active_model/array_serializer'
 require 'active_model/serializable'
 require 'active_model/serializer/associations'
-require 'active_model/serializer/settings'
+require 'active_model/serializer/config'
 
 module ActiveModel
   class Serializer
@@ -14,19 +14,19 @@ module ActiveModel
       end
 
       def setup
-        yield SETTINGS
+        yield CONFIG
       end
 
       def embed(type, options={})
-        SETTINGS[:embed] = type
-        SETTINGS[:include] = true if options[:include]
+        CONFIG.embed = type
+        CONFIG.include = true if options[:include]
         warn <<-WARN
 ** Notice: embed is deprecated. **
 The use of .embed method on a Serializer will be soon removed, as this should have a global scope and not a class scope.
 Please use the global .setup method instead:
 ActiveModel::Serializer.setup do |config|
   config.embed = :#{type}
-  config.include = #{SETTINGS[:include] || false}
+  config.include = #{CONFIG.include || false}
 end
         WARN
       end