+++ lib/active_model/serializer/associations.rb
@@ -8,8 +8,8 @@ module ActiveModel
         @name          = name.to_s
         @options       = options
 
-        self.embed     = options[:embed] || SETTINGS[:embed] || :objects
-        @embed_in_root = @embed_ids && (options[:include] || SETTINGS[:include])
+        self.embed     = options[:embed] || CONFIG.embed || :objects
+        @embed_in_root = @embed_ids && (options[:include] || CONFIG.include)
         @embed_key     = options[:embed_key] || :id
         @key           = options[:key]
         @embedded_key  = options[:root] || name